+++ lib/src/utils/gtk_colors.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart';
+import 'package:xdg_directories/xdg_directories.dart' as xdgdir;
 
 import 'dart:io';
 import 'dart:ui';
@@ -9,6 +10,7 @@ import 'package:gsettings/gsettings.dart';
 class GnomeTheme {
   String contents = "";
 
+  late String theme;
   Color? themeBgColor;
   Color? textColor;
   Color? buttonFg;
@@ -25,10 +27,25 @@ class GnomeTheme {
     parse();
   }
 
-  void loadFromFile() {
+  loadFromFile() {
     final settings = GSettings(schemaId: 'org.gnome.desktop.interface');
-    final theme = settings.stringValue('gtk-theme');
-    String fileName = "/usr/share/themes/" + theme + "/gtk-3.0/gtk.css";
+    theme = settings.stringValue('gtk-theme');
+    late String fileName;
+    String path1 = "/usr/share/themes/" + theme + "/gtk-3.0/gtk.css";
+    String path2 = xdgdir.configHome.path.replaceAll('config', 'themes/') +
+        theme +
+        "/gtk-3.0/gtk.css";
+    String path3 =
+        "${xdgdir.dataHome.path}/themes/" + theme + "/gtk-3.0/gtk.css";
+    if (Directory(path1).existsSync()) {
+      fileName = path1;
+    } else if (Directory(path2).existsSync()) {
+      fileName = path2;
+    } else if (Directory(path3).existsSync()) {
+      fileName = path3;
+    } else {
+      return null;
+    }
     File file = File(fileName);
     if (file.existsSync()) {
       contents = File(fileName).readAsStringSync();
@@ -38,7 +55,11 @@ class GnomeTheme {
   ThemeData data(context) {
     ThemeData themedata = ThemeData(
         iconTheme: IconThemeData(color: textColor),
-        brightness: Brightness.dark,
+        brightness: textColor != null
+            ? textColor!.computeLuminance() <= 0.5
+                ? Brightness.light
+                : Brightness.dark
+            : null,
         canvasColor: themeBgColor,
         textSelectionTheme: TextSelectionThemeData(cursorColor: textColor),
         elevatedButtonTheme: ElevatedButtonThemeData(
@@ -60,14 +81,23 @@ class GnomeTheme {
     RegExpMatch? matches = exp.firstMatch(contents);
     String? color = matches?.group(2).toString();
     if (color != null) {
-      return fromCssColor(color);
+      try {
+        return fromCssColor(color);
+      } catch (e) {
+        debugPrint(e.toString());
+      }
     }
     return null;
   }
 
   void parse() {
     themeBgColor = getBaseColor("theme_base_color");
-    textColor = getBaseColor("theme_text_color");
+    textColor = getBaseColor("theme_text_color") ??
+        (theme.contains('dark') ||
+                (themeBgColor != null &&
+                    themeBgColor!.computeLuminance() <= 0.5)
+            ? Colors.white
+            : Colors.black);
     buttonFg = getBaseColor("button_fg_color");
     themeBaseColor = getBaseColor("theme_base_color");
     insensitiveBgColor = getBaseColor("insensitive_bg_color");
@@ -80,26 +110,27 @@ class GnomeTheme {
   }
 }
 
-getGtkColor({
+Color getGtkColor({
   required GtkColorType colorType,
   required bool isDark,
-  GtkColorTheme colorTheme = GtkColorTheme.adwaita,
+  GtkColorTheme colorTheme = GtkColorTheme.system,
 }) {
   switch (colorTheme) {
     case GtkColorTheme.adwaita:
-      return _adwaita(isDark)[describeEnum(colorType)];
-    case GtkColorTheme.system:
+      return _adwaita(isDark)[describeEnum(colorType)]!;
+    default:
       var theme = GnomeTheme();
       theme.loadFromFile();
       theme.parse();
-      return _system(theme)[describeEnum(colorType)];
+      return _system(theme)[describeEnum(colorType)] ??
+          _adwaita(isDark)[describeEnum(colorType)]!;
   }
 }
 
-getAdaptiveGtkColor(
+Color getAdaptiveGtkColor(
   BuildContext context, {
   required GtkColorType colorType,
-  GtkColorTheme colorTheme = GtkColorTheme.adwaita,
+  GtkColorTheme colorTheme = GtkColorTheme.system,
 }) =>
     getGtkColor(
       colorType: colorType,
@@ -130,42 +161,22 @@ enum GtkColorTheme {
   system,
 }
 
-Map<String, Color> _system(GnomeTheme theme) {
+Map<String, Color?> _system(GnomeTheme theme) {
   return {
-    'windowControlIconColor':
-        theme.buttonFg != null ? theme.buttonFg! : const Color(0xFFF9F8F8),
-    'canvas': theme.themeBgColor != null
-        ? theme.themeBgColor!
-        : const Color(0xFFF9F8F8),
-    'headerBarBackgroundTop': theme.themeBgColor != null
-        ? theme.themeBgColor!
-        : const Color(0xFFF9F8F8),
-    'headerBarBackgroundBottom': theme.themeBgColor != null
-        ? theme.themeBgColor!
-        : const Color(0xFFF9F8F8),
-    'headerBarTopBorder':
-        theme.border != null ? theme.border! : const Color(0xFFF9F8F8),
-    'headerBarBottomBorder':
-        theme.border != null ? theme.border! : const Color(0xFFF9F8F8),
-    'headerButtonBackgroundTop':
-        theme.buttonFg != null ? theme.buttonFg! : const Color(0xFFF9F8F8),
-    'headerButtonBackgroundBottom':
-        theme.buttonFg != null ? theme.buttonFg! : const Color(0xFFF9F8F8),
-    'headerButtonBackgroundTopHover': theme.themeSelectedBG != null
-        ? theme.themeSelectedBG!.withOpacity(0.2)
-        : const Color(0xFFF9F8F8),
-    'headerButtonBackgroundBottomHover': theme.themeSelectedBG != null
-        ? theme.themeSelectedBG!.withOpacity(0.2)
-        : const Color(0xFFF9F8F8),
-    'headerButtonBorder': theme.unfocusedBorder != null
-        ? theme.unfocusedBorder!
-        : const Color(0xFFF9F8F8),
-    'headerButtonPrimary':
-        theme.buttonFg != null ? theme.buttonFg! : const Color(0xFFF9F8F8),
-    'headerSwitcherTabPrimary':
-        theme.buttonFg != null ? theme.buttonFg! : const Color(0xFFF9F8F8),
-    'headerSwitcherTabBackground': const Color(0xFFD5D1CD),
-    'headerSwitcherTabBorder': const Color(0xFFCDC7C2),
+    'windowControlIconColor': theme.buttonFg,
+    'canvas': theme.themeBgColor,
+    'headerBarBackgroundTop': theme.themeBgColor,
+    'headerBarBackgroundBottom': theme.themeBgColor,
+    'headerBarTopBorder': theme.border,
+    'headerBarBottomBorder': theme.border,
+    'headerButtonBackgroundTop': theme.buttonFg,
+    'headerButtonBackgroundBottom': theme.buttonFg,
+    'headerButtonBackgroundTopHover': theme.themeSelectedBG?.withOpacity(0.2),
+    'headerButtonBackgroundBottomHover':
+        theme.themeSelectedBG?.withOpacity(0.2),
+    'headerButtonBorder': theme.unfocusedBorder,
+    'headerButtonPrimary': theme.buttonFg,
+    'headerSwitcherTabPrimary': theme.buttonFg,
   };
 }